02_TOOLSETS/policy_analysis/requirements.txt
# US Flag Vessel Analysis - Dependencies
# Python 3.11+ required (not installed via pip)

# Web scraping and automation
playwright>=1.41.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
xlrd>=2.0.1

# PDF handling
pdfplumber>=0.10.3
reportlab>=4.0.9
PyPDF2>=3.0.1

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.1
plotly>=5.18.0

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging and monitoring
tqdm>=4.66.1
colorama>=0.4.6

# Data validation
pydantic>=2.5.0

# HTTP client improvements
httpx>=0.26.0
aiohttp>=3.9.1

# Date handling
python-dateutil>=2.8.2

# Installation instructions:
# 1. pip install -r requirements.txt
# 2. playwright install chromium
